requirements.txt
mysql-connector-python==8.0.33
matplotlib==3.7.1
python-dotenv==1.0.0
tkcalendar==1.6.1
